#434 merge loqs models
