+++ pybamm/models/lead_acid/__init__.py
@@ -4,6 +4,5 @@
 from .loqs import LOQS
 from .composite import Composite
 from .newman_tiedemann import NewmanTiedemann
-from .loqs_capacitance import LOQSCapacitance
 from .composite_capacitance import CompositeCapacitance
 from .newman_tiedemann_capacitance import NewmanTiedemannCapacitance
