+++ tests/integration/test_models/test_lead_acid/test_loqs_capacitance.py
@@ -10,36 +10,35 @@ import numpy as np
 
 
 class TestLeadAcidLOQSCapacitance(unittest.TestCase):
-    def test_basic_processing(self):
-        model = pybamm.lead_acid.LOQSCapacitance()
+    def test_basic_processing_differential(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.LOQS(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
     @unittest.skipIf(scikits_odes_spec is None, "scikits.odes not installed")
-    def test_basic_processing_no_capacitance(self):
-        model = pybamm.lead_acid.LOQSCapacitance(use_capacitance=False)
+    def test_basic_processing_algebraic(self):
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.LOQS(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all()
 
-    def test_basic_processing_1p1D(self):
-        bc_options = {"dimensionality": 2}
-        # pybamm.set_logging_level("DEBUG")
-        model = pybamm.lead_acid.LOQSCapacitance(bc_options=bc_options)
+    def test_basic_processing_1p1D_differential(self):
+        options = {"capacitance": "differential", "bc_options": {"dimensionality": 2}}
+        model = pybamm.lead_acid.LOQS(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all(skip_output_tests=True)
 
     @unittest.skipIf(scikits_odes_spec is None, "scikits.odes not installed")
-    def test_basic_processing_1p1D_no_capacitance(self):
-        bc_options = {"dimensionality": 2}
-        # pybamm.set_logging_level("DEBUG")
-        model = pybamm.lead_acid.LOQSCapacitance(
-            use_capacitance=False, bc_options=bc_options
-        )
+    def test_basic_processing_1p1D_algebraic(self):
+        options = {"capacitance": "algebraic", "bc_options": {"dimensionality": 2}}
+        model = pybamm.lead_acid.LOQS(options)
         modeltest = tests.StandardModelTest(model)
         modeltest.test_all(skip_output_tests=True)
 
     def test_optimisations(self):
-        model = pybamm.lead_acid.LOQSCapacitance()
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.LOQS(options)
         optimtest = tests.OptimisationsTest(model)
 
         original = optimtest.evaluate_model()
