+++ tests/unit/test_models/test_lead_acid/test_loqs.py
@@ -21,6 +21,39 @@ class TestLeadAcidLOQS(unittest.TestCase):
         self.assertTrue("negative particle" not in model.default_geometry)
 
 
+class TestLeadAcidLOQSCapacitance(unittest.TestCase):
+    def test_well_posed_differential(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.LOQS(options)
+        model.check_well_posedness()
+
+    def test_well_posed_algebraic(self):
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.LOQS(options)
+        model.check_well_posedness()
+
+    def test_well_posed_1plus1D(self):
+        options = {"capacitance": "differential", "bc_options": {"dimensionality": 2}}
+        model = pybamm.lead_acid.LOQS(options)
+        model.check_well_posedness()
+
+    def test_default_geometry(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.LOQS(options)
+        self.assertNotIn("current collector", model.default_geometry)
+        options["dimensionality"] = 2
+        model = pybamm.lead_acid.LOQS(options)
+        self.assertIn("current collector", model.default_geometry)
+
+    def test_default_solver(self):
+        options = {"capacitance": "differential"}
+        model = pybamm.lead_acid.LOQS(options)
+        self.assertIsInstance(model.default_solver, pybamm.ScikitsOdeSolver)
+        options = {"capacitance": "algebraic"}
+        model = pybamm.lead_acid.LOQS(options)
+        self.assertIsInstance(model.default_solver, pybamm.ScikitsDaeSolver)
+
+
 if __name__ == "__main__":
     print("Add -v for more debug output")
     import sys
